examples/list_printers/requirements.txt
cachetools==3.1.0
certifi==2018.11.29
chardet==3.0.4
google-auth==1.6.2
idna==2.8
pyasn1==0.4.5
pyasn1-modules==0.2.4
requests==2.21.0
rsa==4.0
six==1.12.0
urllib3==1.24.1
